+++ spec/rubocop/cop/rspec/hooks_before_examples_spec.rb
@@ -227,4 +227,65 @@ RSpec.describe RuboCop::Cop::RSpec::HooksBeforeExamples do
       RUBY
     end
   end
+
+  context 'when Ruby 3.4', :ruby34 do
+    it 'flags `around` after `it`' do
+      expect_offense(<<~RUBY)
+        RSpec.describe User do
+          it { is_expected.to be_after_around_hook }
+          around { it }
+          ^^^^^^^^^^^^^ Move `around` above the examples in the group.
+        end
+      RUBY
+
+      expect_correction(<<~RUBY)
+        RSpec.describe User do
+          around { it }
+          it { is_expected.to be_after_around_hook }
+        end
+      RUBY
+    end
+
+    it 'flags `around` after `context`' do
+      expect_offense(<<~RUBY)
+        RSpec.describe User do
+          context 'a context' do
+            it { is_expected.to be_after_around_hook }
+          end
+
+          around { it }
+          ^^^^^^^^^^^^^ Move `around` above the examples in the group.
+        end
+      RUBY
+
+      expect_correction(<<~RUBY)
+        RSpec.describe User do
+          around { it }
+          context 'a context' do
+            it { is_expected.to be_after_around_hook }
+          end
+
+        end
+      RUBY
+    end
+
+    it 'flags `around` after `include_examples`' do
+      expect_offense(<<~RUBY)
+        RSpec.describe User do
+          include_examples('should be after around-hook')
+
+          around { it }
+          ^^^^^^^^^^^^^ Move `around` above the examples in the group.
+        end
+      RUBY
+
+      expect_correction(<<~RUBY)
+        RSpec.describe User do
+          around { it }
+          include_examples('should be after around-hook')
+
+        end
+      RUBY
+    end
+  end
 end